+++ src/main/java/com/bookbla/americano/domain/member/service/MemberStyleService.java
@@ -2,7 +2,7 @@ package com.bookbla.americano.domain.member.service;
 
 import com.bookbla.americano.domain.member.Member;
 import com.bookbla.americano.domain.member.MemberStyle;
-import com.bookbla.americano.domain.member.controller.dto.StylesResponse;
+import com.bookbla.americano.domain.member.controller.dto.response.StylesResponse;
 import com.bookbla.americano.domain.member.enums.ContactType;
 import com.bookbla.americano.domain.member.enums.DateCostType;
 import com.bookbla.americano.domain.member.enums.DateStyle;
@@ -11,8 +11,8 @@ import com.bookbla.americano.domain.member.enums.JustFriendType;
 import com.bookbla.americano.domain.member.enums.SmokeType;
 import com.bookbla.americano.domain.member.repository.MemberRepository;
 import com.bookbla.americano.domain.member.repository.MemberStyleRepository;
-import com.bookbla.americano.domain.member.service.dto.MemberStyleResponse;
-import com.bookbla.americano.domain.member.service.dto.MemberStyleUpdateRequest;
+import com.bookbla.americano.domain.member.controller.dto.response.MemberStyleResponse;
+import com.bookbla.americano.domain.member.controller.dto.request.MemberStyleUpdateRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
