+++ src/test/java/com/bookbla/americano/domain/member/service/MemberStyleServiceTest.java
@@ -20,12 +20,12 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import com.bookbla.americano.base.exception.BaseException;
 import com.bookbla.americano.domain.member.Member;
 import com.bookbla.americano.domain.member.MemberStyle;
-import com.bookbla.americano.domain.member.controller.dto.StylesResponse;
+import com.bookbla.americano.domain.member.controller.dto.response.StylesResponse;
 import com.bookbla.americano.domain.member.enums.MemberType;
 import com.bookbla.americano.domain.member.repository.MemberRepository;
 import com.bookbla.americano.domain.member.repository.MemberStyleRepository;
-import com.bookbla.americano.domain.member.service.dto.MemberStyleResponse;
-import com.bookbla.americano.domain.member.service.dto.MemberStyleUpdateRequest;
+import com.bookbla.americano.domain.member.controller.dto.response.MemberStyleResponse;
+import com.bookbla.americano.domain.member.controller.dto.request.MemberStyleUpdateRequest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
